Rename MemcachedCacheEntryHttp to HttpByteArrayCacheEntrySerializer.
